Fixed #92: New "include 2.0" does not update consistently
This time for real I hope ;)
Interface changes:
Method 'render' in KnowWEDomRenderer now got the additional attribute 'KnowWEArticle article'.
This is necessary since the article that is accessible via the method getArticle() in Sections is not necessarily the article this Section is rendered for, because Sections can also be included. Choosing the right article is particularly important for the usage of the SectionStores, because every Section uses separate stores for each article it is included to. The same applies to ReviseSubtreeHandlers. Be aware, that a Section possibly gets parsed in different articles, so it is important to use the article given in the 'reviseSubtree' method for retrieving the right KnowledgeBase since this is the one calling the handler.